+++ src/libs/react-router-dom/auth-route.tsx
@@ -1,5 +1,5 @@
 import { useDialog } from "@/components/dialog/dialog.hook"
-import { PATH_SIGN_IN } from "@/constants"
+import { PATH_SIGN } from "@/constants"
 import { AUTH_TOKEN } from "@/constants/auth-key"
 import { useEffect } from "react"
 import { Outlet, useNavigate } from "react-router-dom"
@@ -12,7 +12,7 @@ export const AuthRoute = () => {
     const accessToken = localStorage.getItem(AUTH_TOKEN)
     if (!!!accessToken) {
       onAlert({ children: "로그인 후, 이용해주세요 :)" })
-      navigate(PATH_SIGN_IN)
+      navigate(PATH_SIGN)
     }
   }, [])
   return <Outlet />
